requirements.txt
numpy
pandas
plotly
Flask
gunicorn
PyYAML
python-dotenv
SQLAlchemy
PyMySQL
Pillow
google-cloud-storage
